dashboard/requirements.txt
django
djangorestframework
django-cors-headers
drf-nested-routers
launchpadlib
celery
bs4
diskcache
dlrnapi_client
jinja2
